requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
model-mommy==2.0.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
toposort==1.7
